mimik_backend/src/modules/requirements.txt
openai_whisper==20230314
TTS==0.12.0